+++ src/js/Model.js
@@ -496,7 +496,7 @@ var Model = {
 						disableInitialConsole: true,
 						use: "HTML5",
 						j2sPath: window.location.origin + window.location.pathname + "src/js/lib/jmol/j2s",
-						script: 'frank off; background black; set antialiasDisplay true; set disablePopupMenu true; set showunitcelldetails false; set hoverDelay 0.001; hover off; set MessageCallback "Model.JSmol.onMessage";',
+						script: 'frank off; background black; unbind "MIDDLE DRAG" "_rotateZorZoom"; bind "MIDDLE DRAG" "_translate"; set antialiasDisplay true; set disablePopupMenu true; set showunitcelldetails false; set hoverDelay 0.001; hover off; set MessageCallback "Model.JSmol.onMessage";',
 						readyFunction: Model.JSmol.onReady.bind(Model.JSmol),
 						console: "none"
 					});